minor refactoring
Integrating SQI in ecg module
